Update gdc.gs
Moved maybeCloseAttrs to the beginning of handleText to make sure that any attribute tags are closed even when there is no additional whitespace between different format types. 

Moved superscript and subscript attribute changes (push) to the rest of the attribute changes to avoid opening a new tag before closing the old ones.